feat: Protection des ressources (Policies)  et Documentation de l'API (Swagger)

+++ app/Http/Controllers/Api/CategoryController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\Plat;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
 
@@ -14,9 +16,12 @@ class CategoryController extends Controller
         $categories = $request->user()->categories()->latest()->get();
         return response()->json($categories);
     }
-    public function show(Request $request, $id)
+
+
+    public function show(Request $request, Category $category)
     {
-        $category = $request->user()->categories()->findOrFail($id);
+        Gate::authorize('view', $category);
+
         return response()->json($category);
     }
 
@@ -40,17 +45,19 @@ class CategoryController extends Controller
         ], 201);
     }
 
-    public function update(Request $request, $id)
+    public function update(Request $request, Category $category)
     {
+        Gate::authorize('update', $category);
+
         $fields = $request->validate([
             'name' => [
                 'required',
                 'max:250',
                 'string',
-                Rule::unique('categories')->where('user_id', $request->user()->id)->ignore($id)
+                Rule::unique('categories')->where('user_id', $request->user()->id)->ignore($category)
             ]
         ]);
-        $category = $request->user()->categories()->findOrFail($id);
+
         $category->update($fields);
 
         return response()->json([
@@ -59,13 +66,45 @@ class CategoryController extends Controller
         ], 200);
     }
 
-    public function destroy(Request $request, $id)
+    public function destroy(Request $request, Category $category)
     {
-        $category = $request->user()->categories()->findOrFail($id);
+
+        Gate::authorize('delete', $category);
+
         $category->delete();
+
         return response()->json([
             'message' => 'Delete successfully',
             'data' => $category
         ], 200);
+    }
+
+    public function associatePlats(Request $request, Category $category)
+    {
+        Gate::authorize('update', $category);
+
+        $request->validate([
+            'plat_ids' => 'required|array',
+            'plat_ids.*' => 'exists:plats,id'
+        ]);
+
+
+        $plats = $request->user()->plats()->whereIn('id', $request->plat_ids)->get();
+
+        $category->plats()->saveMany($plats);
+
+        return response()->json([
+            'message' => 'Plats associés avec succès',
+        ], 200);
+    }
+
+    public function getPlatsByCategory(Category $category)
+    {
+
+        Gate::authorize('view', $category);
+
+        $plats = $category->plats()->latest()->get();
+
+        return response()->json($plats, 200);
     }
 }
